+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 + TEALv5 support in `@algo-builder/runtime` [AVM 1.0](https://developer.algorand.org/articles/discover-avm-10/):
     * Cover, Uncover opcodes
     * Loads, Stores opcodes
+    * Extract, Extract3 opcodes
+    * ExtractUint16, ExtractUint32, ExtractUint64 opcodes
 
 ## v2.1.0 2021-10-22